[CICV-35] conn_test: fix bad memory reference
The variable u16SerBytesCount is 2 bytes but gets refereced as 4 bytes.
In the called function it is assiged corrupting memory around it.

+++ ieee_802_15_4/examples/connectivity_test/connectivity_test.c
@@ -342,7 +342,7 @@ smacErrors_t smacToAppMcpsSap(smacToAppDataMessage_t* pMsg, instanceId_t instanc
 
 static void HandleEvents(int32_t evSignals)
 {
-    uint16_t u16SerBytesCount = 0;
+    uint32_t u16SerBytesCount = 0;
 
     if(evSignals & gUART_RX_EVENT_c)
     {
